training/requirements/requirements_tools.txt
Pillow>=11.0.0
torch
torchvision
